chore: 폴더 구조 변경 components/header 내부에서 controller, view, event 함수를 따로 관리하도록 수정
