+++ index.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="./src/styles/colors.css" />
     <link rel="stylesheet" href="./src/styles/fonts.css" />
     <link rel="stylesheet" href="./src/styles/global.css" />
-    <link rel="stylesheet" href="./src/components/header.css" />
+    <link rel="stylesheet" href="./src/components/header/header.css" />
     <link rel="stylesheet" href="./src/components/transactionform.css" />
     <link rel="stylesheet" href="./src/components/transactionList.css" />
     <title>Document</title>
