+++ src/index.js
@@ -1,5 +1,5 @@
 import { createMainLayout } from "./pages/main.js";
-import { initHeader } from "./components/header.js";
+import { initHeader } from "./components/header/HeaderController.js";
 
 initHeader();
 createMainLayout();
